# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/customer/": {
            "get": {
                "tags": [
                    "Customer"
                ],
                "summary": "Get list of customer",
                "description": "Returns a list of customer",
                "operationId": "customer/list",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer  create",
                "description": "Customer  Customer",
                "operationId": "Customer/store",
                "requestBody": {
                    "description": "Customer create",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "location_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": " name",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": " phone",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": " email",
                                        "type": "string"
                                    },
                                    "location_id": {
                                        "description": " location_id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/customer/{id}": {
            "put": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer  create",
                "description": "Customer  Customer",
                "operationId": "Customer/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id for category",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Customer create",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "location_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": " name",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": " phone",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": " email",
                                        "type": "string"
                                    },
                                    "location_id": {
                                        "description": " location_id",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customer"
                ],
                "summary": "Customer detete",
                "description": "delete  Customer",
                "operationId": "Customer/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pass id for Customer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Failed operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}